requirements.txt
Flask=3.11.9
python-slugify=8.0.4
argon2-cffi=23.1.0